+++ examples/remove_settable.py
@@ -1,8 +1,5 @@
 #!/usr/bin/env python
-"""A sample application for cmd2 demonstrating how to remove one of the built-in runtime settable parameters.
-
-It also demonstrates how to use the cmd2.Cmd.select method.
-"""
+"""A sample application for cmd2 demonstrating how to remove one of the built-in runtime settable parameters."""
 
 import cmd2
 
@@ -11,12 +8,6 @@ class MyApp(cmd2.Cmd):
     def __init__(self) -> None:
         super().__init__()
         self.remove_settable('debug')
-
-    def do_eat(self, arg):
-        sauce = self.select('sweet salty', 'Sauce? ')
-        result = '{food} with {sauce} sauce, yum!'
-        result = result.format(food=arg, sauce=sauce)
-        self.stdout.write(result + '\n')
 
 
 if __name__ == '__main__':
